+++ src/main/java/de/bushnaq/abdalla/kassandra/repository/TaskRepository.java
@@ -31,5 +31,5 @@ public interface TaskRepository extends ListCrudRepository<TaskDAO, Long> {
     List<TaskDAO> findBySprintIdOrderByOrderIdAsc(Long sprintId);
 
     @Query("SELECT COALESCE(MAX(t.orderId), 0) FROM TaskDAO t")
-    Long findMaxOrderId();
+    Integer findMaxOrderId();
 }
